corrected installation instructions and cleaned up setup.py & readme.md
== Installation

* installation of this library was unclear --
** both setup.py and requirements.txt provided, and unconnected to each
other (latter removed)
** unused/unnecessary requirements (commented-out/removed respectively)

* instruction to build & install appears incorrect --
** generally unnecessary to `setup.py build` & `setup.py install`; the
latter command does both
** these instructions insufficient, due to scikit-learn -- setuptools
alone *cannot* install this dependency in this manner, and so this
procedure fails. (rather this dependency, and therefore the entire
library, must be installed via `pip` -- as it would be installed from
PyPI.)

* python version requirement loosened (no need to require a micro
version -- rather, should aim to support the remainder of the major
version 3)

* library likely can't yet be considered "production/stable" (rather,
pre-alpha)

* software license as yet undetermined (but likely not Python
Foundation's own)

== Documentation

readme.md cleaned up for clarity of language and formatting, and to
correct installation instructions.