Remove unnecessary 0s from features (#17)
As of now, we truncate/append 0s so that all rows are of length `dim`. However, we don't update `dim`, but rather `self.dim`, depending on the type of features. This PR makes it so that we do truncation according to the correct value of `dim`, ensuring that extraneous 0s are not added.

Co-authored-by: Jesse London <jesselondon@gmail.com>